Scripting: Contextまわりの整理
