+++ GenerativeArtWorkbench/Presentation/Views/Script/Core/JSImage.swift
